+++ src/main/kotlin/hu/neuron/entity/PostCodeToCity.kt
@@ -5,7 +5,7 @@ import org.springframework.data.annotation.Id
 data class PostCodeToCity (
 
     @Id
-    val postcode: Long = 0,
+    val postcode: Int = 0,
 
     val city: String = ""
